feat: `async` Database and DAO Repositories
The SQLite database can now be accessed through an asynchronous
interface using the "Repository" objects. The Repository objects provide
a relatively clean interface to the database; you do not have to worry
about queries. Example code:

```java
Context context = getApplicationContext();
RecipeRepository recipeRepository = new RecipeRepository(context);
recipeRepository.getAll().observe(this, recipes -> {
			Log.d("DATABASE", String.valueOf(recipes));
});
```

The above initializes an "observer" that will wait for the asynchronous
call to finish and execute the code in the subsequent block. See
AndroidStudio documentation for `LiveData<>` for more. Working with the
`TagRepository` and `RecipeTagJoinRepository` classes is done similarly.

Keep in mind the database is currently _in-memory_, and so no changes to
the database are persistent.